MODE72: New sprite mode added.
This mode tries to imitate the Commodore 64 VIC-II chip's parameters in
multicolor sprite mode. It supports up to 18 x 12 pixels wide sprites of
which 8 may display on a scanline. There is no X mirroring, sprite height
is fixed at 21 pixels, and colors 2 and 3 are locked, taken from Sprite 0
for all sprites.