MODE72: Rewrote sprite system by new concept.
The new concept allows each sprite slot to have any number of sprites by
chaining them in Y order. Also ROM sprites are introduced. Currently most
sprite modes are removed except for the basic 8 px wide sprites.

Note that the demo currently does not work and many things are in
progress. The basic features are tested, it is capable to display
sprites.